+++ packages/terra-framework-docs/src/terra-dev-site/doc/data-grid/WorklistDataGrid.1/Examples.3/WorklistDataGridFocusableCell.jsx
@@ -31,10 +31,10 @@ const WorklistDataGridFocusableCell = () => {
 
   const gridDataJSON = {
     cols: [
-      { id: 'Column-0', displayName: 'Patient' },
-      { id: 'Column-1', displayName: 'Column 1' },
+      { id: 'Column-1', displayName: 'Patient' },
       { id: 'Column-2', displayName: 'Column 2' },
       { id: 'Column-3', displayName: 'Column 3' },
+      { id: 'Column-4', displayName: 'Column 4' },
     ],
     rows: [
       {