+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated `terra-compact-interactive-list` documentation for `terra-menu` support.
+
 ## 1.73.0 - (March 8, 2024)
 
 * Changed